Add workaround for stuttering/crackling in Chrome 89
I have nothing better to do than write a hypothetical explanation.

For some reason, onaudioprocess is no longer being called (or running?)
fast enough to generate samples on time, producing crackles. As a
workaround, generate samples for the next frame while the current is
playing, then just copy the buffer. Even if my hypothesis is wrong, this
appears to fix the problem.

The correct solution is almost certainly to switch to AudioWorklets
instead of the deprecated ScriptProcessorNode.